rag-fastapi/requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
pytest>=8.0,<9.0
httpx>=0.28,<1.0
numpy>=1.26,<3.0
requests>=2.32,<3.0
sentence-transformers==5.1.2
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
passlib>=1.7,<2.0
python-dotenv>=1.0,<2.0
python-jose[cryptography]>=3.3,<4.0
python-multipart>=0.0.9,<1.0
